+++ core/properties/__init__.py
@@ -1,8 +1,17 @@
 from core.properties.common import FluidTransportProperties
 from core.properties.fluids import ConstantPropertyProvider, PropertyProvider
 
+from core.properties.water import (
+    IAPWS97WaterSteamProvider,
+    WaterSteamProperties,
+    water_steam_props_iapws97,
+)
+
 __all__ = [
     "FluidTransportProperties",
     "PropertyProvider",
     "ConstantPropertyProvider",
+    "IAPWS97WaterSteamProvider",
+    "WaterSteamProperties",
+    "water_steam_props_iapws97",
 ]